requirements.txt
alabaster==1.0.0
babel==2.17.0
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
defusedxml==0.7.1
docutils==0.21.2
dropbox==11.36.2
Flask==3.0.0
idna==3.10
imagesize==1.4.1
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
pdfservices-sdk==4.2.0
Pygments==2.19.2
python-dotenv==1.1.1
requests==2.32.5
snowballstemmer==3.0.1
Sphinx==8.1.3
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
tomli==2.2.1
urllib3==2.5.0
Werkzeug==3.0.1
